Added z parameter to crystal_map
Moved z/depth functionality from separate crystal_map_3d class to crystal_map. Deleted crystal_map_3d.py
